lading page + announcement v1
backend:
store announcementu,
announcement controller, middleware pre checkrole pre admin / garant
laravel password pre API autentifikáciu

frontend:
pinia store na správu stavu
quill.js editor
zobrazenie oznámení
role-based UI v dashboard view
jednotný navbar v app.vue
prepísanie tailwind na bootstrap classes + odstránenie tailwind komponentov
fix yellow warnings

+++ backend/database/migrations/2025_10_11_134947_create_companies_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->integer('id', true);
             $table->string('name', 100);
             $table->integer('address_id')->nullable()->index('address_id');
-            $table->integer('user_id')->nullable()->index('companies_user_fk');
+            $table->unsignedBigInteger('user_id')->nullable()->index('companies_user_fk');
         });
     }
